requirements.txt
aiogram==3.0.0rc2
aiohttp==3.8.5
APScheduler==3.10.0
gunicorn==20.1.0
Jinja2==3.1.2
python-dotenv==0.21.0
redis==4.5.1
requests==2.28.1
tqdm==4.64.1
webdriver-manager==3.8.5
pytz~=2022.1